devops_agent/requirements.txt
open-interpreter
dotenv
